+++ src/us/kbase/workspace/test/kbase/LoggingTest.java
@@ -829,6 +829,18 @@ public class LoggingTest {
 				new AdminExp("end method", SERV))));
 		logout.reset();
 		
+		// get objects
+		ac.put("command", "getObjects");
+		ac.put("params", new GetObjects2Params()
+				.withObjects(Arrays.asList(new ObjectSpecification().withRef("1/1/1"))));
+		CLIENT2.administer(new UObject(ac));
+		checkLogging(convertAdminExp(Arrays.asList(
+				new AdminExp("start method", SERV),
+				new AdminExp("getObjects", ADMIN),
+				new AdminExp("Object 1/1/1 SomeModule.AType-1.0", ARGUTILS),
+				new AdminExp("end method", SERV))));
+		logout.reset();
+		
 		// list ws
 		ac.put("command", "listWorkspaces");
 		ac.put("user", USER1);